Refactored to move route logic to individual files
